refactor: remove unwated eslint disable rules
