+++ README.md
@@ -1,2 +1,2 @@
 [![Build Status](https://img.shields.io/travis/rachellesarkis1/js-stack.svg?style=flat-square)](https://travis-ci.org/rachellesarkis1/js-stack)
-[![Coverage Status](https://img.shields.io/coveralls/rachellesarkis1/js-stack.svg?style=flat-square)](https://coveralls.io/github/rachellesarkis1/js-stack?branch=master)
+[![Coverage Status](https://img.shields.io/coveralls/rachellesarkis1/js-stack.svg?style=flat-square)](https://coveralls.io/github/rachellesarkis1/js-stack?branch=main)
