+++ src/client/index.jsx
@@ -1,5 +1,4 @@
 // @flow
-/* eslint quotes: ["error", "double"] */
 
 import Immutable from "immutable";
 
@@ -9,22 +8,20 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { AppContainer } from "react-hot-loader";
 import { Provider } from "react-redux";
-/* eslint-disable */
 import { createStore, combineReducers, applyMiddleware, compose } from "redux";
 import thunkMiddleware from "redux-thunk";
+import { BrowserRouter } from "react-router-dom";
 import setUpSocket from "./socket";
 
 import App from "../shared/app";
 import helloReducer from "../shared/reducer/hello";
-import { APP_CONTAINER_SELECTOR } from "../shared/config";
+import { APP_CONTAINER_SELECTOR, JSS_SSR_SELECTOR } from "../shared/config";
 import { isProd } from "../shared/util";
-import { BrowserRouter } from "react-router-dom";
-import { JSS_SSR_SELECTOR } from "../shared/config";
 
 window.jQuery = $;
 window.Tether = Tether;
 require("bootstrap");
-// eslint-disable-next-line no-underscore-dangle
+/* eslint-disable no-underscore-dangle */
 const composeEnhancers =
   (isProd ? null : window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__) || compose;
 const preloadedState = window.__PRELOADED_STATE__;
