+++ src/server/render-app.jsx
@@ -1,6 +1,4 @@
 // @flow
-/* eslint quotes: ["error", "double"] */
-/* eslint-disable */
 
 import { SheetsRegistry, JssProvider } from "react-jss";
