+++ src/shared/app.jsx
@@ -1,5 +1,4 @@
 // @flow
-/* eslint quotes: ["error", "double"] */
 
 import React from "react";
 import { Switch } from "react-router";
