+++ src/shared/component/message.jsx
@@ -1,5 +1,4 @@
 // @flow
-/* eslint quotes: ["error", "double"] */
 
 import React from "react";
