+++ src/shared/component/page/hello.jsx
@@ -1,5 +1,4 @@
 // @flow
-/* eslint quotes: ["error", "double"] */
 
 import React from "react";
 import Helmet from "react-helmet";
