+++ src/shared/config.js
@@ -1,7 +1,5 @@
 // @flow
 
-/* eslint quotes: ["error", "double"] */
-
 export const WEB_PORT = process.env.PORT || 8000;
 export const STATIC_PATH = "/static";
 export const APP_NAME = "Hello World";
