+++ src/shared/container/hello-button.js
@@ -1,5 +1,4 @@
 // @flow
-/* eslint quotes: ["error", "double"] */
 
 import { connect } from "react-redux";
