+++ src/shared/routes.js
@@ -1,8 +1,5 @@
 // @flow
 
-// eslint-disable-next-line import/prefer-default-export
-/* eslint-disable */
-
 export const HOME_PAGE_ROUTE = "/";
 export const HELLO_PAGE_ROUTE = "/hello";
 export const HELLO_ASYNC_PAGE_ROUTE = "/hello-async";
