+++ src/shared/util.js
@@ -1,5 +1,4 @@
 // @flow
 
-/* eslint quotes: ["error", "double"] */
 // eslint-disable-next-line import/prefer-default-export
 export const isProd = process.env.NODE_ENV === "production";
